+++ app/src/main/java/com/woowa/banchan/ui/orderdetail/OrderDetailViewModel.kt
@@ -16,37 +16,12 @@ import javax.inject.Inject
 
 @HiltViewModel
 class OrderDetailViewModel @Inject constructor(
-    private val getOrderInfoUseCase: GetOrderInfoUseCase,
     private val getOrderLineItemUseCase: GetOrderLineItemUseCase,
     private val modifyOrderUseCase: ModifyOrderUseCase
 ) : ViewModel() {
 
-    private val _orderState = MutableStateFlow(OrderUiState())
-    val orderState = _orderState.asStateFlow()
-
     private val _state = MutableStateFlow(OrderLineItemUiState())
     val state = _state.asStateFlow()
-
-    fun getAllOrder() {
-        viewModelScope.launch {
-            getOrderInfoUseCase().onEach { result ->
-                result.onSuccess {
-                    _orderState.value = _orderState.value.copy(
-                        orderInfoList = it,
-                        isLoading = false,
-                        errorMessage = ""
-                    )
-                }
-                    .onFailure {
-                        _state.value = _state.value.copy(
-                            orderLineItemList = mapOf(),
-                            isLoading = false,
-                            errorMessage = "주문 내역 불러오기가 실패했습니다."
-                        )
-                    }
-            }.launchIn(this)
-        }
-    }
 
     fun getOrderLineItem(orderId: Int) {
         viewModelScope.launch {
